+++ src/main/java/moze_intel/projecte/gameObjs/items/tools/DarkHoe.java
@@ -23,6 +23,7 @@ public class DarkHoe extends PEToolBase
 	@Override
 	public boolean onItemUse(ItemStack stack, EntityPlayer player, World world, int x, int y, int z, int par7, float par8, float par9, float par10)
 	{
-		return tillSoil(stack, player, world, x, y, z, par7);
+		tillSoil(stack, player, world, x, y, z, par7);
+		return true;
 	}
 }
