+++ src/main/java/moze_intel/projecte/gameObjs/items/tools/DarkPick.java
@@ -1,30 +1,14 @@
 package moze_intel.projecte.gameObjs.items.tools;
 
 import moze_intel.projecte.gameObjs.ObjHandler;
-import moze_intel.projecte.gameObjs.entity.EntityLootBall;
-import moze_intel.projecte.network.PacketHandler;
-import moze_intel.projecte.network.packets.SwingItemPKT;
 import moze_intel.projecte.utils.AchievementHandler;
-import moze_intel.projecte.utils.Coordinates;
-import moze_intel.projecte.utils.Utils;
 import net.minecraft.block.Block;
 import net.minecraft.block.material.Material;
 import net.minecraft.entity.EntityLivingBase;
 import net.minecraft.entity.player.EntityPlayer;
-import net.minecraft.entity.player.EntityPlayerMP;
-import net.minecraft.init.Blocks;
 import net.minecraft.item.ItemStack;
-import net.minecraft.util.AxisAlignedBB;
-import net.minecraft.util.MathHelper;
-import net.minecraft.util.MovingObjectPosition;
-import net.minecraft.util.MovingObjectPosition.MovingObjectType;
 import net.minecraft.util.StatCollector;
 import net.minecraft.world.World;
-import net.minecraftforge.common.ForgeHooks;
-import net.minecraftforge.common.util.ForgeDirection;
-
-import java.util.ArrayList;
-import java.util.List;
 
 public class DarkPick extends PEToolBase
 {
@@ -50,135 +34,10 @@ public class DarkPick extends PEToolBase
 	@Override
 	public boolean onBlockDestroyed(ItemStack stack, World world, Block block, int x, int y, int z, EntityLivingBase eLiving)
 	{
-		if (world.isRemote || !(eLiving instanceof EntityPlayer))
-		{
-			return false;
-		}
-		
-		EntityPlayer player = (EntityPlayer) eLiving;
-		byte mode = this.getMode(stack);
-		
-		if (mode == 0)
-		{
-			return false;
-		}
-			
-		MovingObjectPosition mop = this.getMovingObjectPositionFromPlayer(world, player, false);
-		AxisAlignedBB box = null;
-		
-		if (mop == null || mop.typeOfHit != MovingObjectType.BLOCK)
-		{
-			return false;
-		}
-		
-		ForgeDirection direction = ForgeDirection.getOrientation(mop.sideHit);
-		
-		if (mode == 1)
-		{
-			box = AxisAlignedBB.getBoundingBox(x, y - 1, z, x, y + 1, z);
-		}
-		else if (mode == 2)
-		{
-			if (direction.offsetX != 0)
-			{
-				box = AxisAlignedBB.getBoundingBox(x, y, z - 1, x, y, z + 1);
-			}
-			else if (direction.offsetZ != 0)
-			{
-				box = AxisAlignedBB.getBoundingBox(x - 1, y, z, x + 1, y, z);
-			}
-			else
-			{
-				int dir = MathHelper.floor_double(player.rotationYaw * 4.0F / 360.0F + 0.5D) & 3;
-				
-				if (dir == 0 || dir == 2)
-				{
-					box = AxisAlignedBB.getBoundingBox(x, y, z - 1, x, y, z + 1);
-				}
-				else
-				{
-					box = AxisAlignedBB.getBoundingBox(x - 1, y, z, x + 1, y, z);
-				}
-			}
-		}
-		else
-		{
-			if (direction.offsetX == 1)
-			{
-				box = AxisAlignedBB.getBoundingBox(x - 2, y, z, x, y, z);
-			}
-			else if (direction.offsetX == - 1)
-			{
-				box = AxisAlignedBB.getBoundingBox(x, y, z, x + 2, y, z);
-			}
-			else if (direction.offsetZ == 1)
-			{
-				box = AxisAlignedBB.getBoundingBox(x, y, z - 2, x, y, z);
-			}
-			else if (direction.offsetZ == -1)
-			{
-				box = AxisAlignedBB.getBoundingBox(x, y, z, x, y, z + 2);
-			}
-			else if (direction.offsetY == 1)
-			{
-				box = AxisAlignedBB.getBoundingBox(x, y - 2, z, x, y, z);
-			}
-			else 
-			{
-				box = AxisAlignedBB.getBoundingBox(x, y, z, x, y + 2, z);
-			}
-		}
-		
-		List<ItemStack> drops = new ArrayList<ItemStack>();
-		
-		for (int i = (int) box.minX; i <= box.maxX; i++)
-			for (int j = (int) box.minY; j <= box.maxY; j++)
-				for (int k = (int) box.minZ; k <= box.maxZ; k++)
-				{
-					Block b = world.getBlock(i, j, k);
-					
-					if (b != Blocks.air && b.getBlockHardness(world, i, j, k) != -1 && (canHarvestBlock(block, stack) || ForgeHooks.canToolHarvestBlock(block, world.getBlockMetadata(i, j, k), stack)))
-					{
-						drops.addAll(Utils.getBlockDrops(world, player, b, stack, i, j, k));
-						world.setBlockToAir(i, j, k);
-					}
-				}
-			
-		world.spawnEntityInWorld(new EntityLootBall(world, drops, x, y, z));
-		return false;
+		digBasedOnMode(stack, world, block, x, y, z, eLiving);
+		return true;
 	}
-	
-	@Override
-	public ItemStack onItemRightClick(ItemStack stack, World world, EntityPlayer player)
-	{
-		if (!world.isRemote)
-		{
-			int offset = this.getCharge(stack) + 3;
-			AxisAlignedBB box = AxisAlignedBB.getBoundingBox(player.posX - offset, player.posY - offset, player.posZ - offset, player.posX + offset, player.posY + offset, player.posZ + offset);
-			List<ItemStack> drops = new ArrayList();
-			
-			for (int x = (int) box.minX; x <= box.maxX; x++)
-				for (int y = (int) box.minY; y <= box.maxY; y++)
-					for (int z = (int) box.minZ; z <= box.maxZ; z++)
-					{
-						Block block = world.getBlock(x, y, z);
-						
-						if (Utils.isOre(block) && block.getBlockHardness(world, x, y, z) != -1 && (canHarvestBlock(block, stack) || ForgeHooks.canToolHarvestBlock(block, world.getBlockMetadata(x, y, z), stack)))
-						{
-							Utils.harvestVein(world, player, stack, new Coordinates(x, y, z), block, drops, 0);
-						}
-					}
-			
-			if (!drops.isEmpty())
-			{
-				world.spawnEntityInWorld(new EntityLootBall(world, drops, player.posX, player.posY, player.posZ));
-				PacketHandler.sendTo(new SwingItemPKT(), (EntityPlayerMP) player);
-			}
-		}
-		
-		return stack;
-	}
-	
+
 	@Override
 	public float getDigSpeed(ItemStack stack, Block block, int metadata)
 	{
